+++ StaticFields/Tests/tests.h
@@ -1,5 +1,5 @@
 #ifndef TESTS_H
 #define TESTS_H
-#include "../src/Control/FieldParser.h"
+#include "../headers/Control/FieldParser.h"
 void CSVTest();
 #endif
