+++ StaticFields/src/Model/Vector.c
@@ -1,5 +1,5 @@
-#include "Vector.h"
-#include "Field.h"
+#include "../../headers/Model/Vector.h"
+#include "../../headers/Model/Field.h"
 #include <stdio.h>
 #include <stdlib.h>
 Vector vecSum(Vector a, Vector b){
